Moved admin navigation guards into a single hook
no ref

The member detail screen had to compose two guards — `useBlocker` for
react-router navigations and `useHashLinkNavigationGuard` for native
`<a href="#/…">` anchors — and hand-roll the blocked/proceed/reset
branching between them. Which path a navigation took is the guard's
business, not the screen's, so `useNavigationGuard` now covers both and
exposes one surface. Dirtiness is passed as a callback rather than a
value so a screen can waive the guard for its own post-save redirect
without waiting on a re-render.
