+++ apps/admin/src/members/detail/member-detail.tsx
@@ -17,10 +17,9 @@ import {formatMemberName} from '@tryghost/shade/app';
 import {getMember, useAddMember, useEditMember} from '@tryghost/admin-x-framework/api/members';
 import {getSettingValue, useBrowseSettings} from '@tryghost/admin-x-framework/api/settings';
 import {toast} from 'sonner';
-import {useBlocker} from 'react-router';
 import {useBrowseNewsletters} from '@tryghost/admin-x-framework/api/newsletters';
 import {useBrowseTiers} from '@tryghost/admin-x-framework/api/tiers';
-import {useHashLinkNavigationGuard} from '@/hooks/use-hash-link-navigation-guard';
+import {useNavigationGuard} from '@/hooks/use-navigation-guard';
 import type {MemberEditableFields} from './member-detail-edit';
 
 // The create screen reuses this route with the sentinel id "new". Safe because
@@ -275,14 +274,10 @@ const MemberDetailPage: React.FC<MemberDetailPageProps> = ({paidMembersEnabled, 
     };
 
     useConfirmUnload(activeMutation.isPending || hasUnsavedChanges);
-    const blocker = useBlocker(({currentLocation, nextLocation}) => !bypassGuardRef.current && hasUnsavedChanges && currentLocation.pathname !== nextLocation.pathname);
-    // Native `<a href="#/…">` navigations (the sidebar, links into Ember
-    // routes) never reach the react-router blocker above — see the hook.
-    // Ember's own guard (`trailing-hash.js`) can't cover this screen either:
-    // with `memberDetailsReact` on, the Ember member route aborts and never
-    // registers into the `unsaved-changes` service.
-    const anchorGuard = useHashLinkNavigationGuard(hasUnsavedChanges);
-    const isBlocked = blocker.state === 'blocked' || anchorGuard.isBlocked;
+    // Ember's `unsaved-changes` service can't guard this screen: with
+    // `memberDetailsReact` on, the Ember member route aborts and never
+    // registers into it.
+    const guard = useNavigationGuard(() => !bypassGuardRef.current && hasUnsavedChanges);
 
     // Save button state (Save / Saving / Saved / Retry), mirroring the Ember task button.
     let saveLabel: React.ReactNode = 'Save';
@@ -452,10 +447,9 @@ const MemberDetailPage: React.FC<MemberDetailPageProps> = ({paidMembersEnabled, 
                     )}
                 </DetailPage.Body>
 
-                <AlertDialog open={isBlocked} onOpenChange={(open) => {
-                    if (!open && isBlocked) {
-                        blocker.reset?.();
-                        anchorGuard.reset();
+                <AlertDialog open={guard.isBlocked} onOpenChange={(open) => {
+                    if (!open) {
+                        guard.reset();
                     }
                 }}>
                     <AlertDialogContent>
@@ -465,13 +459,7 @@ const MemberDetailPage: React.FC<MemberDetailPageProps> = ({paidMembersEnabled, 
                         </AlertDialogHeader>
                         <AlertDialogFooter>
                             <AlertDialogCancel>Keep editing</AlertDialogCancel>
-                            <Button variant='destructive' onClick={() => {
-                                if (anchorGuard.isBlocked) {
-                                    anchorGuard.proceed();
-                                } else {
-                                    blocker.proceed?.();
-                                }
-                            }}>Leave</Button>
+                            <Button variant='destructive' onClick={() => guard.proceed()}>Leave</Button>
                         </AlertDialogFooter>
                     </AlertDialogContent>
                 </AlertDialog>
